k线大屏/quantify-master/backend/requirements.txt
# Web框架
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# 数据处理
pandas>=2.0.0
numpy>=1.24.0
tushare>=1.2.89
setuptools>=65.0.0

# 数据库
sqlalchemy==2.0.23
sqlmodel==0.0.14
alembic==1.12.1
pymysql==1.1.0
cryptography==41.0.7

# 任务调度
apscheduler==3.10.4
croniter==1.4.1
cron-descriptor==1.2.24
pytz==2023.3

# HTTP客户端
httpx==0.25.2
requests==2.31.0

# 配置管理
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# 日志
loguru==0.7.2

# JWT认证
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# 开发工具
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-xdist==3.5.0
black==23.11.0
isort==5.12.0
flake8==6.1.0

# 缓存（可选）
redis==5.0.1

# 系统监控
psutil==5.9.6

# 同花顺自选依赖
browser-cookie3==0.19.1

# 同花顺登录依赖
ddddocr>=1.4.0
PyExecJS>=1.5.1
playwright>=1.40.0
